pyproject.toml
[project]
name = "france-healthcare-access"
version = "0.2.0"
description = "Multi-source ELT pipeline + BI dashboard mapping GP accessibility across France."
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "streamlit>=1.35.0",
    "pandas>=2.2.2",
    "plotly>=5.22.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.5",
    "pyarrow>=17.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.8",
    "black>=24.4.2",
]

[tool.ruff]
line-length = 100
